Fabric 1.17
It's time

Method got its params changed from FDD to DDD, very weird.
Everything else works the same despite being in Java 16.
What was that flamingo thing about anyway? Must have been part of a fabric example somewhere.

+++ src/main/java/com/charles445/damagetilt/KnockbackHandler.java
@@ -8,7 +8,7 @@ import net.minecraft.server.network.ServerPlayerEntity;
 
 public class KnockbackHandler
 {
-	public static void onKnockback(Object entity, float f, double d, double e)
+	public static void onKnockback(Object entity, double f, double d, double e)
 	{
 		if(entity instanceof PlayerEntity)
 		{
